1.2.0(3)

+++ lib/ui/widgets/actions/rec_del.dart
@@ -7,6 +7,8 @@ import 'package:everlong/services/recorder.dart';
 import 'package:everlong/ui/widgets/progress_indicator.dart';
 import 'package:everlong/ui/widgets/button.dart';
 import 'package:everlong/ui/widgets/svg.dart';
+import 'package:everlong/ui/widgets/confirm.dart';
+import 'package:everlong/ui/widgets/dialog.dart';
 import 'package:everlong/ui/widgets/snackbar.dart';
 import 'package:everlong/utils/constants.dart';
 import 'package:everlong/utils/texts.dart';
@@ -29,11 +31,21 @@ class RecordDelete extends StatelessWidget {
       isActive: false,
       icon: svgIcon(name: kRecDelIcon, color: color, width: kIconWidth),
       onPressed: () {
-        if (fileType == FileType.recording) {
-          context.read<Recorder>().clear();
-        } else if (fileType == FileType.stored) {
-          context.read<Recorder>().deleteRecord(id!);
-        }
+        showDialog(
+          context: context,
+          builder: (BuildContext context) => dialogBox(
+            smallerDialog: true,
+            context: context,
+            content: confirmDialog(context, onConfirm: () async {
+              if (fileType == FileType.recording) {
+                context.read<Recorder>().clear();
+              } else if (fileType == FileType.stored) {
+                context.read<Recorder>().deleteRecord(id!);
+              }
+              Navigator.pop(context);
+            }),
+          ),
+        );
       },
     );
   }
